+++ source/Rtsi/RtsiClient.cpp
@@ -3,6 +3,7 @@
 #include "RtsiRecipeInternal.hpp"
 #include "Utils.hpp"
 #include "VersionInfo.hpp"
+#include "Log.hpp"
 
 #include <array>
 #include <iostream>
@@ -29,6 +30,9 @@ void RtsiClient::connect(const std::string& ip, int port) {
         socket_ptr_->async_connect(endpoint, [&](const boost::system::error_code& error) {
             if (!error) {
                 connection_state = ConnectionState::CONNECTED;
+            } else {
+                connection_state = ConnectionState::DISCONNECTED;
+                ELITE_LOG_ERROR("Connect to RTSI server %s:%d fail: %s", ip.c_str(), port, error.message().c_str());
             }
         });
         io_context_.run();
@@ -192,6 +196,7 @@ void RtsiClient::sendAll(const PackageType& cmd, const std::vector<uint8_t>& pay
     boost::system::error_code ec;
     socket_ptr_->write_some(boost::asio::buffer(message), ec);
     if (ec) {
+        ELITE_LOG_FATAL("RTSI socket send fail: %s", ec.message().c_str());
         throw EliteException(EliteException::Code::SOCKET_FAIL, ec.message());
     }
 }
@@ -209,6 +214,7 @@ int RtsiClient::receiveSocket(std::vector<uint8_t>& buff, int size, int offset, 
     boost::asio::async_read(*socket_ptr_, boost::asio::buffer(buff.data() + offset, size),
                             [&](const boost::system::error_code& ec, std::size_t nb) {
                                 if (ec) {
+                                    ELITE_LOG_FATAL("RTSI socket receive fail: %s", ec.message().c_str());
                                     throw EliteException(EliteException::Code::SOCKET_FAIL, ec.message());
                                 }
                                 read_len = nb;
@@ -224,7 +230,7 @@ int RtsiClient::receiveSocket(std::vector<uint8_t>& buff, int size, int offset, 
     // the pending asynchronous operation is a composed operation, the deadline
     // applies to the entire operation, rather than individual operations on
     // the socket.
-    io_context_.run_for(std::chrono::steady_clock::duration(std::chrono::milliseconds(timeout_ms)));
+    io_context_.run_for(std::chrono::milliseconds(timeout_ms));
 
     // If the asynchronous operation completed successfully then the io_context
     // would have been stopped due to running out of work. If it was not